+++ mock-abis/src/main/java/io/mosip/proxy/abis/dao/ProxyAbisInsertRepository.java
@@ -1,9 +1,9 @@
-package com.proxy.abis.dao;
+package io.mosip.proxy.abis.dao;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import com.proxy.abis.entity.InsertEntity;
+import io.mosip.proxy.abis.entity.InsertEntity;
 
 @Repository
 public interface ProxyAbisInsertRepository extends CrudRepository<InsertEntity, String>{
