+++ mock-abis/src/main/java/io/mosip/proxy/abis/entity/ResponseMO.java
@@ -1,4 +1,4 @@
-package com.proxy.abis.entity;
+package io.mosip.proxy.abis.entity;
 
 import java.time.LocalDateTime;
